pyproject.toml
[project]
name = "phylogen"
version = "0.1.0"
description = "Transformer model for valid bacterial mutations by leveraging phylogenetic relationships."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "torch",
    "biopython",
    "matplotlib",
    "seaborn",
    "pandas",
    "tqdm"
]
